backend/requirements.txt
python-decouple==3.8
python-multipart==0.0.6 
requests==2.28.2
fastapi==0.92.0
tweepy
langchain
openai
chromadb
plawyright
dotenv-python
google-search-results